added KloudlitePodActiveLabel for first successful ping in pod-pinger

+++ operators/networking/internal/cmd/ip-binding-controller/pod-pinger/controller.go
@@ -33,6 +33,8 @@ type Reconciler struct {
 	yamlClient kubectl.YAMLClient
 }
 
+const KloudlitePodActiveLabel = "kloudlite.io/pod.active"
+
 func (r *Reconciler) GetName() string {
 	return r.Name
 }
@@ -85,10 +87,20 @@ func (r *Reconciler) Reconcile(ctx context.Context, request ctrl.Request) (ctrl.
 	pinger.Timeout = 500 * time.Millisecond
 	if err = pinger.RunWithContext(ctx); err != nil {
 		r.logger.Errorf(err, "failed to ping %s", v)
+		if _, ok := pod.Labels[KloudlitePodActiveLabel]; ok {
+			return ctrl.Result{RequeueAfter: 5 * time.Second}, nil
+		}
 		if err := r.Delete(ctx, pod); err != nil {
 			return ctrl.Result{}, err
 		}
 		return ctrl.Result{}, err
+	}
+	if _, ok := pod.Labels[KloudlitePodActiveLabel]; !ok {
+		pod.Labels[KloudlitePodActiveLabel] = "true"
+		if err := r.Update(ctx, pod); err != nil {
+			r.logger.Errorf(err, "failed to update pod %s/%s", pod.GetNamespace(), pod.GetName())
+			return ctrl.Result{}, err
+		}
 	}
 
 	r.logger.Debugf("ping success for %s, requeing after 5s", v)
